backend/requirements.txt
# FastAPI and web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Stock data and analysis
yfinance>=0.2.32
pandas>=2.1.3

# Configuration and environment
python-dotenv>=1.0.0

# HTTP requests
requests>=2.31.0

# Task scheduling
schedule>=1.2.0

# Email (built-in smtplib is used, no extra package needed)
